server/requirements.txt
langchain==0.3.7
langchain-core==0.3.17
langchain-community==0.3.5
langchain-text-splitters==0.3.2
langchain-anthropic==0.3.0
langchain-huggingface==0.1.2
anthropic>=0.73.0
faiss-cpu==1.8.0
sentence-transformers==3.3.1
pymupdf==1.24.0
python-docx==1.1.0
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.0
pydantic==2.12.4
nltk==3.9.2
